pyproject.toml
[tool.poetry]
name = "osukoreanbot"
version = "0.1.0"
description = "Discord bot for OSU Korean Discord Server"
authors = ["OSUCoding"]
license = "GNU 3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyNaCl = "^1.5.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
